Adaptive Antyaliasing done, changed class Sample, refactoring sample forcebrute

+++ src/main/java/fgk/zad1/basics/Vector2.java
@@ -4,7 +4,7 @@ public class Vector2 {
     public float x;
     public float y;
 
-    public Vector2(int x, int y) {
+    public Vector2(float x, float y) {
         this.x = x;
         this.y = y;
     }
